refactor: consolidate EventRouter into ForemanWss; extract assignment into TaskManager
- Delete `EventRouter` class; routing logic becomes free functions in
  `event-routing.ts` (`routeEvent`, `routePrEvent`, `routeCheckEvent`,
  etc.) taking an explicit `RoutingDeps` object — no constructor callbacks
- Move `isMutedEvent`, `summaryEvent` to `event-fmt.ts` (already the home
  for event formatting helpers); remove the only other import site that
  used `event-router.ts` (http-server.ts)
- Add `TaskManager.reserveTaskForWorker(workerId)`: atomic find-and-assign
  with an internal `assignLock` mutex, replacing the lock that lived in
  `wss.ts`; `ForemanWss` calls this instead of the former `tryAssignWork`
- `ForemanWss` retains `sendMsg` as a local 5-line helper (protocol concern)
  and constructs a `RoutingDeps` object to pass to routing free functions
- Update all tests to import from new locations; rewrite handler tests to
  call free functions directly with a `RoutingDeps` object; add tests for
  `TaskManager.reserveTaskForWorker` (null/success/mutex/error propagation)
- Update CLAUDE.md to document the new structure

Closes #653
