+++ src/foreman/event-fmt.ts
@@ -80,3 +80,55 @@ export function fmtEvent(event: { name: string; payload: Record<string, unknown>
   const details = fmtEventDetails(event);
   return `${nameAction}${details ? ` — ${details}` : ""}`;
 }
+
+// ── Webhook event summary (one-line log output) ───────────────────────────────
+
+function truncTitle(title: unknown, max = 50): string {
+  const s = String(title ?? "").replace(/\s+/g, " ").trim();
+  return s.length > max ? s.slice(0, max - 1) + "…" : s;
+}
+
+export function summaryEvent(id: string, name: string, payload: unknown): string {
+  const p = payload as Record<string, unknown>;
+  const action = typeof p.action === "string" ? `/${p.action}` : "";
+  const repo = (p.repository as Record<string, unknown> | undefined)?.full_name;
+  const sender = (p.sender as Record<string, unknown> | undefined)?.login;
+
+  let detail = "";
+  const issue = p.issue as Record<string, unknown> | undefined;
+  const pr = p.pull_request as Record<string, unknown> | undefined;
+
+  if (issue) {
+    detail = ` #${issue.number} "${truncTitle(issue.title)}"`;
+  } else if (pr) {
+    detail = ` #${pr.number} "${truncTitle(pr.title)}"`;
+  } else if (name === "push") {
+    const ref = String(p.ref ?? "");
+    const count = (p.commits as unknown[] | undefined)?.length ?? 0;
+    detail = ` ${ref} (${count} commit${count === 1 ? "" : "s"})`;
+  } else if (name === "delete") {
+    const ref = String(p.ref ?? "");
+    if (ref) detail = ` ${ref}`;
+  } else if (name === "check_run" || name === "check_suite" || name === "workflow_run" || name === "workflow_job") {
+    const inner = p[name] as Record<string, unknown> | undefined;
+    const prs = inner?.pull_requests as Array<{ number: number }> | undefined;
+    if (prs && prs.length > 0) {
+      detail = ` PR #${prs[0].number}`;
+    } else {
+      const headBranch = name === "check_run"
+        ? String((inner?.check_suite as Record<string, unknown> | undefined)?.head_branch ?? "")
+        : String(inner?.head_branch ?? "");
+      if (headBranch) detail = ` ${headBranch}`;
+    }
+  }
+
+  const parts: string[] = [`${name}${action}${detail}`];
+  if (sender) parts.push(`by ${sender}`);
+  if (repo) parts.push(`(${repo})`);
+
+  return `[event] ${parts.join(" ")}`;
+}
+
+export function isMutedEvent(name: string): boolean {
+  return name === "workflow_job" || name === "workflow_run";
+}
